backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.0
fpdf==1.7.2
Werkzeug==2.3.7
